feat: chalk package done
