+++ packages/chalk/src/index.ts
@@ -1,3 +1,2 @@
-import { chalk } from './ansi-colors.js'
-
-console.log(chalk.greenBright().bgBlue().bold().italic('HELLO'))
+export * from './sgr.js'
+export * from './chalk.js'
